Unit Test Refactor (#497)
* Fixed EPERM Windows Testing Errors

This fixes EPERM errors thrown by klaw-sync. For some reason klaw works fines so we can just use klaw instead.

* Roosevelt test refactor

This refactors roosevelt.js tests to fix most of the issues we are having with them on Windows.

* Move test certs into their own directory

This closes #377 by moving test certs into their own directory and also refactors them in the roosevelt test to be a reusable object instead of multiple variables.

* Replace node_modules with mocks

This replaces node_modules in unit tests with mocks so we have complete control over them and there are no false negatives in the future. Closes #384

* Updated package

Updated package-lock / package for build testing.

* Removed commented https test

* Update unit test from pervious merge

* Remove .only

Remove .only from test.

* Updates to roosevelt test

This should fix some of the build issues and inconsistent passes and fails.

* Downgraded eslint

Downgraded eslint for issues with the build.

* Test App Generator Cleanup

This cleanup just makes this a bit easier to read and more structured for easier additions / deletions down the road.

* Reintroduce latest eslint

The linter issue seems to be fixed so this updates eslint back to the current version.

* Fixes package-lock for the linter

* Added stopServer feature

This will add a roosevelt object member stopServer() with an optional argument of 'close' that can be passed.

* Refactor multi-thread tests

This refactors the multi-thread tests to reflect the new roosevelt stopServer() feature.

* Windows Unit Test Refactor

All tests will now pass on windows and also report accurate line coverage.

* Simplify stopServer

* Clean up joined paths

This just cleans up path.join() from having many arguments.

* Removed two unnecessary tests

* Copy editing, comments, small refactors

This just copy edits, comments, and refactors many pieces of the tests.

* Fix the two broken tests

* Refactor for appveyor

* Updated travis and appveyor

* Auto-kill test regression

This regression is to the previous state of one of the auto-kill tests which seems to make the build fail on travis.

* Remove node 9 from builds

This removes node 9 from travis and appveyor. We will only keep the Node stable and Node latest.

* eslint downgrade

This downgrades eslint to 4.19.1 until the issues with 5 are resolved.