Added support for suppressing environments
This will add support for a 'disable' param to be set in package.json that takes an array of items. That array will take in different environments or anything process.env key that is set to true ex: ['production', 'LOADED_MOCHA_OPTS'] will suppress production logs and console logs in mocha that shouldn't be there.