Winston Logger (#512)
* Winston logger initial commit

This takes the original roosevelt logger and replaces it with the winston logger. This also exposes the logger and the winston logger module via app.get('logger'). To access winston you simply require the logger and then use logger.winston.

* Allow for custom logging parameters

This will allow for custom logging params to create custom log types ex: logger.custom('Hello World')

* remove semver from winston dependency

* Fix proper emoji spacing break

When using emoji.which() with node-emoji. If an object is passed in, it currently breaks so this adds a typeof check before.

* Fix for config auditor test

This commit changes the way we read the config auditor output data. Winston uses process.stdout.write & process.stderr.write vs our old logger which used console.log & console.error. With that being said we need to hook into the stream to push its data into an array.

* Check to make sure that the type is a string before we check if it's an emoji

* warning log level change

By default winston outputs to the standard output stream, even if the level is `warn`.  So this just adds the level 'warn' to the stderr stream instead.

* Added unit test for the logger

* Fixed the autoKillTest

* Logger changes

This adds more consistent string formatting, gives transports its own object to easily override settings, and fixes the default logtype issues.

* More consistent spacing

* More consistent spacing

This gives more consistent spacing to the logger for the first argument if it's an emoji.

* Remove .only

* copyedit comments

* Added support for suppressing environments

This will add support for a 'disable' param to be set in package.json that takes an array of items. That array will take in different environments or anything process.env key that is set to true ex: ['production', 'LOADED_MOCHA_OPTS'] will suppress production logs and console logs in mocha that shouldn't be there.

* Clean up test

logger unit test cleanup

* small refactor on string spacing

* Suppress logs in the params test

* remove unnecessary check

* trim whitespace from the end of strings

* cleanup comment

* README update

* Fix for the config auditor

This ignores the extra keys check for the logging object

* Fix for enable param

This fixes the enable parameter to work properly if set to false

* Revert the kill signal log

* Consistent spacing with the kill signal log

* Change enabled to enable for consistency

* README update

* Add tests for disabled logs

Added some extra tests for disabled logs to better test the loggers functionality

* Updated package.json

* fresh package-lock

* Linted against standard