Controller files now accept a router and an app object
This commit causes breaking changes in the way Roosevelt controller files are coded. From now on controllers will accept two arguments. The first argument is an express router and the second argument is the app object. Express routes can be written the same way as before but to access anything inside of app you must use the optional second argument.